Update README and fix all clippy warnings
- Update README with new features: WOZ v2, cassette I/O, drag-and-drop,
  clipboard, screenshot, test suite, disk format table
- Fix clippy warnings: unnecessary casts, collapsible ifs, manual div_ceil,
  needless range loops, no-effect operations, Default::default patterns

+++ crates/apple2-audio/src/speaker.rs
@@ -123,7 +123,7 @@ impl Speaker {
 
                 // Accumulate the fraction of this sample at the current level.
                 let frac = (tc - seg_start) / cycles_per_sample;
-                accumulated += frac as f64 * self.level() as f64;
+                accumulated += frac * self.level() as f64;
 
                 // Perform the toggle.
                 self.state = !self.state;
@@ -134,7 +134,7 @@ impl Speaker {
             // Remaining fraction of the sample after the last toggle (or the
             // entire sample if there were no toggles).
             let frac = (sample_end - seg_start) / cycles_per_sample;
-            accumulated += frac as f64 * self.level() as f64;
+            accumulated += frac * self.level() as f64;
 
             let raw = accumulated as f32;
